Make product_generation module
Move notifications into seperate module for namespacing
