+++ shakecast/app/orm/__init__.py
@@ -3,3 +3,6 @@ from .engine import engine
 from .session import Session
 from .migrations import migrate
 from .utils import *
+from .data import load_data
+
+load_data()
