+++ shakecast/app/product_generation/sc_csv.py
@@ -2,7 +2,7 @@ import csv
 import json
 import os
 
-from .notifications import TemplateManager
+from ..notifications.templates import TemplateManager
 
 def generate_impact_csv(shakemap, save=False, file_name='', template_name=''):
     '''
@@ -47,3 +47,6 @@ def save_csv(csv_rows, file_name, directory):
         csv_writer = csv.writer(csv_file)
         for row in csv_rows:
             csv_writer.writerow(row)
+
+def main(*args, **kwargs):
+    generate_impact_csv(*args, **kwargs)
